+++ src/HealthBar.cpp
@@ -34,7 +34,7 @@ void HealthBar::render() {
  * 
  * @damage value in range [0.0f, 1.0f]
  */
-void HealthBar::updateHealth(float damage) {
+void HealthBar::updateHealth(double damage) {
     _health += damage;
     _healthBar->setProgress(_health);
 //    _health->addTween(ProgressBar::TweenProgress(num), 1000);
